Refactor and simplify datasource load
